Move SpawnLevel objects to a spawnlevel package.
Remove NSpawnLevel.
