+++ src/com/palmergames/bukkit/towny/object/spawnlevel/NationSpawnLevel.java
@@ -1,4 +1,4 @@
-package com.palmergames.bukkit.towny.object;
+package com.palmergames.bukkit.towny.object.spawnlevel;
 
 import com.palmergames.bukkit.towny.TownyUniverse;
 import org.bukkit.entity.Player;
@@ -6,6 +6,8 @@ import org.bukkit.entity.Player;
 import com.palmergames.bukkit.config.ConfigNodes;
 import com.palmergames.bukkit.towny.TownySettings;
 import com.palmergames.bukkit.towny.exceptions.TownyException;
+import com.palmergames.bukkit.towny.object.Nation;
+import com.palmergames.bukkit.towny.object.Translation;
 import com.palmergames.bukkit.towny.permissions.PermissionNodes;
 
 public enum NationSpawnLevel {
@@ -59,11 +61,11 @@ public enum NationSpawnLevel {
 
 		if (!isAllowed(player, nation)) {
 			boolean war = nation.hasActiveWar();
-			NSpawnLevel level = TownySettings.getNSpawnLevel(this.isAllowingConfigNode);
-			if(level == NSpawnLevel.WAR && !war) {
+			SpawnLevel level = TownySettings.getSpawnLevel(this.isAllowingConfigNode);
+			if(level == SpawnLevel.WAR && !war) {
 				throw new TownyException(Translation.of(notAllowedLangNodeWar));
 			}
-			else if(level == NSpawnLevel.PEACE && war) {
+			else if(level == SpawnLevel.PEACE && war) {
 				throw new TownyException(Translation.of(notAllowedLangNodePeace));
 			}
 			throw new TownyException(Translation.of(notAllowedLangNode));
@@ -77,8 +79,8 @@ public enum NationSpawnLevel {
 	
 	private boolean isAllowedNation(Nation nation) {
 		boolean war = nation.hasActiveWar();
-		NSpawnLevel level = TownySettings.getNSpawnLevel(this.isAllowingConfigNode);
-		return level == NSpawnLevel.TRUE || (level != NSpawnLevel.FALSE && ((level == NSpawnLevel.WAR) == war));
+		SpawnLevel level = TownySettings.getSpawnLevel(this.isAllowingConfigNode);
+		return level == SpawnLevel.TRUE || (level != SpawnLevel.FALSE && ((level == SpawnLevel.WAR) == war));
 	}
 
 	public double getCost() {
@@ -89,12 +91,5 @@ public enum NationSpawnLevel {
 	public double getCost(Nation nation) {
 
 		return this == NationSpawnLevel.ADMIN ? 0 : nation.getSpawnCost();
-	}
-	
-	public enum NSpawnLevel {
-		TRUE,
-		FALSE,
-		WAR,
-		PEACE
 	}
 }
