+++ src/com/palmergames/bukkit/towny/object/spawnlevel/SpawnLevel.java
@@ -1,0 +1,8 @@
+package com.palmergames.bukkit.towny.object.spawnlevel;
+
+public enum SpawnLevel {
+	TRUE,
+	FALSE,
+	WAR,
+	PEACE
+}
